move stuff into version submodules
