+++ bioimageio/spec/__init__.py
@@ -1,4 +1,9 @@
-from . import nodes
-from .utils import load_and_resolve_spec, load_model_spec, load_spec, maybe_convert_manifest, maybe_convert_model
+from bioimageio.spec.common_utils import CommonVersionSpecificUtils
+from . import v0_1, v0_3
+from .latest import fields, maybe_convert_model, nodes, raw_nodes, schema, utils
 
 __version__ = nodes.FormatVersion.__args__[-1]
+
+load_and_resolve_spec = utils.load_and_resolve_spec
+load_model_spec = utils.load_model_spec
+load_spec = utils.load_spec
