+++ bioimageio/spec/common_utils/common.py
@@ -1,0 +1,12 @@
+from ruamel.yaml import YAML
+
+yaml = YAML(typ="safe")
+
+
+class Singleton(type):
+    _instances = {}
+
+    def __call__(cls, *args, **kwargs):
+        if cls not in cls._instances:
+            cls._instances[cls] = super(Singleton, cls).__call__(*args, **kwargs)
+        return cls._instances[cls]
