+++ bioimageio/spec/cummulative.py
@@ -1,0 +1,5 @@
+import typing
+
+from . import v0_1, v0_3
+
+FormatVersion = typing.Literal[v0_1.FormatVersion, v0_3.FormatVersion]
