+++ bioimageio/spec/latest/__init__.py
@@ -1,0 +1,1 @@
+from bioimageio.spec.v0_3 import fields, maybe_convert_model, nodes, raw_nodes, schema, utils  # noqa
