--- bioimageio/spec/utils/common.py
@@ -1,3 +1,0 @@
-from ruamel.yaml import YAML
-
-yaml = YAML(typ="safe")
