+++ bioimageio/spec/v0_1/raw_nodes.py
@@ -1,0 +1,3 @@
+from bioimageio.spec.v0_3.raw_nodes import *
+
+FormatVersion = Literal["0.1.0"]
