+++ bioimageio/spec/v0_3/__init__.py
@@ -1,0 +1,6 @@
+from bioimageio.spec.common_utils import CommonVersionSpecificUtils
+from . import fields, nodes, raw_nodes, schema
+from ._maybe_convert import maybe_convert_model
+from .raw_nodes import FormatVersion
+
+utils = CommonVersionSpecificUtils(FormatVersion.__args__[-1])
