+++ bioimageio/spec/v0_3/nodes.py
