+++ tests/test_build_spec.py
@@ -1,12 +1,12 @@
 import os
 from pathlib import Path
+
 from bioimageio.spec import maybe_convert_model, schema
-from bioimageio.spec.utils import yaml
+from bioimageio.spec.common_utils import yaml
+from bioimageio.spec.latest.build_spec import _get_local_path, build_spec
 
 
 def test_build_spec_pickle(rf_config_path):
-    from bioimageio.spec.utils.build_spec import build_spec
-
     source = yaml.load(rf_config_path)
     source = maybe_convert_model(source)
 
@@ -48,8 +48,6 @@ def test_build_spec_pickle(rf_config_path):
 
 
 def test_build_spec_pytorch(UNet2DNucleiBroad_model_url):
-    from bioimageio.spec.utils.build_spec import build_spec, _get_local_path
-
     config_path = _get_local_path(UNet2DNucleiBroad_model_url)
     assert os.path.exists(config_path), config_path
     source = yaml.load(Path(config_path))
@@ -86,8 +84,6 @@ def test_build_spec_pytorch(UNet2DNucleiBroad_model_url):
 
 
 def test_build_spec_onnx(UNet2DNucleiBroad_model_url):
-    from bioimageio.spec.utils.build_spec import build_spec, _get_local_path
-
     config_path = _get_local_path(UNet2DNucleiBroad_model_url)
     assert os.path.exists(config_path), config_path
     source = yaml.load(Path(config_path))
@@ -125,8 +121,6 @@ def test_build_spec_onnx(UNet2DNucleiBroad_model_url):
 
 
 def test_build_spec_torchscript(UNet2DNucleiBroad_model_url):
-    from bioimageio.spec.utils.build_spec import build_spec, _get_local_path
-
     config_path = _get_local_path(UNet2DNucleiBroad_model_url)
     assert os.path.exists(config_path), config_path
     source = yaml.load(Path(config_path))
@@ -164,8 +158,6 @@ def test_build_spec_torchscript(UNet2DNucleiBroad_model_url):
 
 
 def test_build_spec_keras(FruNet_model_url):
-    from bioimageio.spec.utils.build_spec import build_spec, _get_local_path
-
     config_path = _get_local_path(FruNet_model_url)
     assert os.path.exists(config_path), config_path
     source = yaml.load(Path(config_path))
@@ -196,8 +188,6 @@ def test_build_spec_keras(FruNet_model_url):
 
 
 def test_build_spec_tf(FruNet_model_url):
-    from bioimageio.spec.utils.build_spec import build_spec, _get_local_path
-
     config_path = _get_local_path(FruNet_model_url)
     assert os.path.exists(config_path), config_path
     source = yaml.load(Path(config_path))
@@ -228,8 +218,6 @@ def test_build_spec_tf(FruNet_model_url):
 
 
 def test_build_spec_tfjs(FruNet_model_url):
-    from bioimageio.spec.utils.build_spec import build_spec, _get_local_path
-
     config_path = _get_local_path(FruNet_model_url)
     assert os.path.exists(config_path), config_path
     source = yaml.load(Path(config_path))
